+++ apps/kyb-app/src/components/organisms/DynamicUI/hooks/useUIStateLogic/hooks/useUIElementsStateLogic/types.ts
@@ -4,6 +4,7 @@ export interface UIElementState {
   isLoading?: boolean;
   isCompleted?: boolean;
   isTouched?: boolean;
+  isDisabled?: boolean;
 }
 
 export interface UIElementsState {