+++ apps/kyb-app/src/components/organisms/UIRenderer/elements/JSONForm/JSONForm.tsx
@@ -7,6 +7,7 @@ import {
   jsonFormFields,
   jsonFormLayouts,
 } from '@app/components/organisms/UIRenderer/elements/JSONForm/json-form.fields';
+import { JSONFormDefinitionProvider } from '@app/components/organisms/UIRenderer/elements/JSONForm/providers/JSONFormDefinitionProvider';
 import { useDataInsertionLogic } from '@app/components/organisms/UIRenderer/hooks/useDataInsertionLogic';
 import { DefinitionInsertionParams } from '@app/components/organisms/UIRenderer/hooks/useDataInsertionLogic/types';
 import { useUIElementErrors } from '@app/components/organisms/UIRenderer/hooks/useUIElementErrors/useUIElementErrors';
@@ -30,11 +31,16 @@ export interface JSONFormElementBaseParams {
 }
 
 export interface JSONFormElementParams extends DefinitionInsertionParams {
-  jsonFormDefinition?: { type?: string; required?: string[] };
+  jsonFormDefinition?: { type?: string; required?: string[]; title?: string };
+  uiSchema?: AnyObject;
 }
 
 export const JSONForm: UIElementComponent<JSONFormElementParams> = ({ definition }) => {
-  useDataInsertionLogic(definition, definition?.options?.jsonFormDefinition?.type !== 'array');
+  useDataInsertionLogic(
+    definition,
+    definition?.options?.jsonFormDefinition?.type !== 'array' ||
+      !definition?.options?.insertionParams,
+  );
 
   const { state: elementState } = useUIElementState(definition);
   const { hidden } = useUIElementProps(definition);
@@ -90,22 +96,26 @@ export const JSONForm: UIElementComponent<JSONFormElementParams> = ({ definition
 
   const { validationErrors } = useUIElementErrors(definition);
 
-  return hidden ? null : (
-    <div className="flex flex-col gap-2">
-      <DynamicForm
-        schema={formSchema}
-        uiSchema={uiSchema}
-        fields={jsonFormFields}
-        layouts={jsonFormLayouts}
-        formData={formData}
-        ref={formRef}
-        transformErrors={transformRJSFErrors}
-        onChange={handleArrayInputChange}
-        onSubmit={handleSubmit}
-      />
-      {validationErrors && elementState.isTouched ? (
-        <ErrorsList errors={validationErrors.map(err => err.message)} />
-      ) : null}
-    </div>
+  return (
+    <JSONFormDefinitionProvider definition={definition}>
+      {hidden ? null : (
+        <div className="flex flex-col gap-2">
+          <DynamicForm
+            schema={formSchema}
+            uiSchema={uiSchema}
+            fields={jsonFormFields}
+            layouts={jsonFormLayouts}
+            formData={formData}
+            ref={formRef}
+            transformErrors={transformRJSFErrors}
+            onChange={handleArrayInputChange}
+            onSubmit={handleSubmit}
+          />
+          {validationErrors && elementState.isTouched ? (
+            <ErrorsList errors={validationErrors.map(err => err.message)} />
+          ) : null}
+        </div>
+      )}
+    </JSONFormDefinitionProvider>
   );
 };