+++ apps/kyb-app/src/components/organisms/UIRenderer/elements/JSONForm/components/DocumentField/DocumentField.tsx
@@ -1,14 +1,14 @@
+import { useEventEmitterLogic } from '@app/components/organisms/DynamicUI/StateManager/components/ActionsHandler';
 import { useStateManagerContext } from '@app/components/organisms/DynamicUI/StateManager/components/StateProvider';
 import { useUIElementToolsLogic } from '@app/components/organisms/DynamicUI/hooks/useUIStateLogic/hooks/useUIElementsStateLogic/hooks/useUIElementToolsLogic';
 import { ErrorField } from '@app/components/organisms/DynamicUI/rule-engines';
 import { getDocumentFileIdPath } from '@app/components/organisms/UIRenderer/elements/JSONForm/components/DocumentField/helpers/getDocumentFileIdPath';
-import { getDocumentFileTypePath } from '@app/components/organisms/UIRenderer/elements/JSONForm/components/DocumentField/helpers/getDocumentFileTypePath';
 import { useUIElementErrors } from '@app/components/organisms/UIRenderer/hooks/useUIElementErrors/useUIElementErrors';
 import { useUIElementState } from '@app/components/organisms/UIRenderer/hooks/useUIElementState';
 import { Document, UIElement } from '@app/domains/collection-flow';
 import { fetchFile, uploadFile } from '@app/domains/storage/storage.api';
 import { collectionFlowFileStorage } from '@app/pages/CollectionFlow/collection-flow.file-storage';
-import { ErrorsList, FileInputAdapter, RJSFInputProps } from '@ballerine/ui';
+import { AnyObject, ErrorsList, FileInputAdapter, RJSFInputProps } from '@ballerine/ui';
 import { HTTPError } from 'ky';
 import get from 'lodash/get';
 import set from 'lodash/set';
@@ -30,15 +30,17 @@ export const DocumentField = (
 
   const { toggleElementLoading } = useUIElementToolsLogic(definition.name);
   const { state: elementState } = useUIElementState(definition);
-  const valueDestination = `document-error-${definition.options.documentData.id}`;
 
   const documentDefinition = useMemo(
     () => ({
       ...definition,
-      valueDestination,
+      valueDestination: `document-error-${definition.options.documentData.id}`,
     }),
     [definition],
   );
+
+  const sendEvent = useEventEmitterLogic(definition);
+
   const { validationErrors, warnings } = useUIElementErrors(documentDefinition);
   const { isTouched } = elementState;
 
@@ -46,7 +48,7 @@ export const DocumentField = (
     if (!Array.isArray(payload.documents)) return null;
 
     const document = payload.documents.find((document: Document) => {
-      return document.id === definition.options.documentData.id;
+      return document?.id === definition.options.documentData.id;
     }) as Document;
 
     const fileIdPath = getDocumentFileIdPath(definition);
@@ -90,7 +92,6 @@ export const DocumentField = (
       }
 
       const fileIdPath = getDocumentFileIdPath(definition);
-      // const fileTypePath = getDocumentFileTypePath(definition);
 
       try {
         const uploadResult = await uploadFile({ file });
@@ -104,24 +105,26 @@ export const DocumentField = (
         collectionFlowFileStorage.registerFile(uploadResult.id, file);
         setFile(file);
         setFieldError(null);
+        sendEvent('onChange');
       } catch (err) {
         if (err instanceof HTTPError) {
-          const jsonError = await err.response.json();
-
+          const response = (await err.response.json()) as AnyObject;
           setFieldError({
             fieldId: document.id,
-            message: jsonError.message,
+            message: response.message as string,
             type: 'warning',
           });
+          return;
         }
-      }
 
-      toggleElementLoading();
+        throw err;
+      } finally {
+        toggleElementLoading();
+      }
     },
-    [stateApi, options, definition, toggleElementLoading],
+    [stateApi, options, definition, toggleElementLoading, sendEvent],
   );
 
-  const filedErrors = fieldError ? [fieldError] : [];
   return (
     <div className="flex flex-col gap-2">
       <FileInputAdapter
@@ -130,9 +133,7 @@ export const DocumentField = (
         formData={fileItem}
         onChange={handleChange}
       />
-      {warnings.length ? (
-        <ErrorsList errors={[...warnings, ...filedErrors].map(err => err.message)} />
-      ) : null}
+      {warnings.length ? <ErrorsList errors={warnings.map(err => err.message)} /> : null}
       {isTouched ? <ErrorsList errors={validationErrors.map(error => error.message)} /> : null}
       {fieldError ? <ErrorsList errors={[fieldError.message]} /> : null}
     </div>