+++ apps/kyb-app/src/components/organisms/UIRenderer/hooks/useDataInsertionLogic/types.ts
@@ -1,3 +1,4 @@
+import { DisablableListElementDefinition } from '@app/components/organisms/UIRenderer/hooks/useDataInsertionLogic/useElementsDisablerLogic';
 import { Rule } from '@app/domains/collection-flow';
 
 export type InsertionSchema = Record<string, string>;
@@ -8,6 +9,8 @@ export interface InsertionParams {
   removeWhen: Rule[];
   insertionStrategy: 'array' | 'object';
   destination: string;
+  disableElements?: DisablableListElementDefinition[];
+  bindingAnchorDestination: string;
 }
 
 export interface DefinitionInsertionParams {