+++ services/workflows-service/src/ui-definition/utils/collection-flow/4-directors-and-ubos.ts
@@ -284,6 +284,7 @@ export const DirectorsAndUbosPage = {
             description:
               '<p>add all the natural persons that own or control, <bold>Directly Or Indirectly</bold> more than 25% of the company.</p>',
             jsonFormDefinition: {
+              title: 'Shareholder',
               type: 'array',
               required: [
                 'ubos:first-name-input',
@@ -295,6 +296,9 @@ export const DirectorsAndUbosPage = {
                 'ubos:address-of-residence-input',
               ],
             },
+            uiSchema: {
+              titleTemplate: 'UBO {{INDEX}}',
+            },
             insertionParams: {
               insertionStrategy: 'array',
               destination: 'entity.data.additionalInfo.ubos',
@@ -303,6 +307,21 @@ export const DirectorsAndUbosPage = {
                 lastName: 'entity.data.additionalInfo.mainRepresentative.lastName',
                 email: 'entity.data.additionalInfo.mainRepresentative.email',
               },
+              bindingAnchorDestination: 'additionalInfo.__isGeneratedAutomatically',
+              disableElements: [
+                {
+                  elementName: 'ubos:first-name-input',
+                  atIndex: 0,
+                },
+                {
+                  elementName: 'ubos:last-name-input',
+                  atIndex: 0,
+                },
+                {
+                  elementName: 'ubos:email-input',
+                  atIndex: 0,
+                },
+              ],
               insertWhen: [
                 {
                   type: 'json-logic',
@@ -331,7 +350,7 @@ export const DirectorsAndUbosPage = {
               type: 'json-form:text',
               valueDestination: 'entity.data.additionalInfo.ubos[{INDEX}].firstName', //entity.data.additionalInfo.ubos[0].firstName
               options: {
-                label: 'Legal Name',
+                label: 'First Name',
                 hint: 'First Name',
                 jsonFormDefinition: {
                   type: 'string',
@@ -480,6 +499,9 @@ export const DirectorsAndUbosPage = {
                     'directors:address-of-residence-input',
                   ],
                 },
+                uiSchema: {
+                  titleTemplate: 'Director {{INDEX}}',
+                },
                 insertionParams: {
                   insertionStrategy: 'array',
                   destination: 'entity.data.additionalInfo.directors',
@@ -488,6 +510,21 @@ export const DirectorsAndUbosPage = {
                     lastName: 'entity.data.additionalInfo.mainRepresentative.lastName',
                     email: 'entity.data.additionalInfo.mainRepresentative.email',
                   },
+                  bindingAnchorDestination: 'additionalInfo.__isGeneratedAutomatically',
+                  disableElements: [
+                    {
+                      elementName: 'directors:first-name-input',
+                      atIndex: 0,
+                    },
+                    {
+                      elementName: 'directors:last-name-input',
+                      atIndex: 0,
+                    },
+                    {
+                      elementName: 'directors:email-input',
+                      atIndex: 0,
+                    },
+                  ],
                   insertWhen: [
                     {
                       type: 'json-logic',
@@ -519,7 +556,7 @@ export const DirectorsAndUbosPage = {
                     jsonFormDefinition: {
                       type: 'string',
                     },
-                    label: 'Legal Name',
+                    label: 'First Name',
                     hint: 'First Name',
                   },
                 },